+++ payment_service/src/main/java/com/easyride/payment_service/dto/WithdrawalRequestDto.java
@@ -9,7 +9,8 @@ public class WithdrawalRequestDto {
 
     private Long driverId;
 
-    private Double amount;
+    // 修改为 Integer 类型，单位为最小货币单位，例如 100 分表示 1 元
+    private Integer amount;
 
     private String bankAccount;
 }
